% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_module_level_analyses.R
\name{analyseClusters}
\alias{analyseClusters}
\title{analyseClusters}
\usage{
analyseClusters(
  CenTFinderObject,
  GO = T,
  KEGG = T,
  PCA = T,
  RCisTarget = T,
  species = c("Homo sapiens", "Mus musculus"),
  cores = detectCores() - 1,
  motif_db_path = NULL
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{GO}{A logical indicating whether Gene Ontology analysis should be performed.}

\item{KEGG}{A logical indicating whether KEGG pathway enrichment should be performed.}

\item{PCA}{A logical indicating whether principal component analysis should be performed.}

\item{RCisTarget}{A logical indicating whether transcription factor binding motif enrichment analysis should be performed. This function is a wrapper for usage of the RCisTarget package.}

\item{species}{Species of the samples. Only Latin names are supported (e.g. Homo sapiens, Mus musculus, ...)}

\item{cores}{Number of cores to be used for RCisTarget analysis. Defaults to total number of cores minus one.}

\item{motif_db_path}{Path to the motif db needed for RCisTarget analysis. For more info, please check http://bioconductor.org/packages/release/bioc/vignettes/RcisTarget/inst/doc/RcisTarget.html.}
}
\value{
An updated CenTFinderObject with cluster analyses.
}
\description{
analyseClusters
}
\examples{
motif_db_path<-"hg19-tss-centered-10kb-7species.mc9nr.feather"
CenTFinderObject<-analyseClusters(CenTFinderObject, GO=T, KEGG=T, PCA=T, RCisTarget=T,species="Homo sapiens", cores=10, motif_db_path=motif_db_path)
}
