% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_WGCNA.R
\name{applyWGCNA}
\alias{applyWGCNA}
\title{applyWGCNA}
\usage{
applyWGCNA(
  CenTFinderObject,
  network_type = c("signed", "unsigned"),
  powers = 1:30,
  cut_height = 0.3
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{network_type}{This argument specifies whether WGCNA should make use of signed or unsigned networks.}

\item{powers}{An integer vector of powers to be evaluated in order to obtain a scale-free topology.}

\item{cut_height}{A numeric specifying the cutting height of the gene tree in order to cluster genes.}
}
\value{
A CenTFinderObject with added TOM, power, connectivity plots, gene cluster annotation, gene tree, and kMEs.
}
\description{
applyWGCNA
}
\examples{
CenTFinderObject<-applyWGCNA(CenTFinderObject, network_type="signed")
}
