% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CenTFinder-class}
\alias{CenTFinder-class}
\alias{CenTFinder}
\title{An S4 class to represent a CenTFinder analysis}
\description{
An S4 class to represent a CenTFinder analysis
}
\section{Slots}{

\describe{
\item{\code{data}}{A gene expression matrix with gene symbols in the rows and samples in the columns.}

\item{\code{sample_annotation}}{Annotation data for each sample (including cell type, array, and accession code.)}

\item{\code{platforms}}{Platform usage information.}

\item{\code{probe_annotations}}{List of additionally provided probe annotations.}

\item{\code{gene_cluster_annotation}}{Gene-to-cluster annotation.}

\item{\code{DE}}{A list of differential expression analyses, each element corresponding to one of the cell types.}

\item{\code{additional_DE}}{Additional differential expression data for the cell type of interest.}

\item{\code{plots}}{Plots generated by WGCNA analysis.}

\item{\code{power}}{The power used for WGCNA soft-tresholding.}

\item{\code{TOM}}{WGCNA topological overlap matrix.}

\item{\code{tree}}{Gene tree generated by hierarchical clustering.}

\item{\code{module_colors}}{Module colour of each gene.}

\item{\code{kMEs}}{Module eigengenes for each gene and each module.}

\item{\code{GO}}{List of Gene Ontology analysis results of each module.}

\item{\code{KEGG}}{List of KEGG enrichment analysis results of each module.}

\item{\code{motif_enrichment}}{Results of the RCisTarget analyses.}

\item{\code{integration}}{List containing an integrative summary of the different analyses, module and regulon scores, and the regulons formatted for Cytoscape visualisation.}
}}

