% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_WGCNA.R
\name{changeModuleColours}
\alias{changeModuleColours}
\title{Change module colours}
\usage{
changeModuleColours(CenTFinderObject, colours = NULL)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{colours}{Vector containing the new module colours. The first colour should be grey, the subsequent colours should be ranked by decreasing module size.}
}
\value{
A CenTFinder object with updated colour scheme.
}
\description{
\code{changeModuleColours} allows to change the cluster colours after WGCNA analysis.
}
