% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_integration.R
\name{compareCTF}
\alias{compareCTF}
\title{compareCTF}
\usage{
compareCTF(ctf1, ctf2, module_annot1, module_annot2)
}
\arguments{
\item{ctf1}{An instance of the CenTFinder class.}

\item{ctf2}{An instance of the CenTFinder class.}

\item{module_annot1}{List with module colour-to-name annotations.}

\item{module_annot2}{List with module colour-to-name annotations.}
}
\value{
An alluvial plot displaying the module memberships. Following CenTFinder releases may extend this comparison.
}
\description{
Compare two CenTFinder analyses
}
\details{
\code{compareCTF} allows one to compare two separate CenTFinder analyses.
}
