% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_WGCNA.R
\name{cutDendrogram}
\alias{cutDendrogram}
\title{Cut WGCNA tree}
\usage{
cutDendrogram(CenTFinderObject, cut_height = 0.2)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{cut_height}{Numeric treshold to cut the dendrogram. Default to 0.2.}
}
\value{
A CenTFinder object with a dendrogram and defined modules.
}
\description{
\code{cutDendrogram} allows custom cutting of the WGCNA dendrogram, as to set the number of clusters.
}
