% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{DE}
\alias{DE}
\title{DE}
\usage{
DE(CenTFinderObject, celltype)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class}

\item{celltype}{Cell type for which the differential expression analysis should be extracted.}
}
\value{
A dataframe with the differential expression results
}
\description{
DE
}
