% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{DEset<-}
\alias{DEset<-}
\title{DEset<-}
\usage{
DEset(x) <- value
}
\arguments{
\item{x}{An instance of the CenTFinder class}

\item{value}{A list containing one or more differential expression analyses, with 'Gene', 'FC', and 'Pvalue' columns in each of the dataframes.}
}
\value{
The CenTFinder object updated with a differential expression set. P-values are automatically Benjamini-Hochberg adjusted.
}
\description{
Load an additional differential expression dataset
}
\details{
\code{'DEset<-'} allows one to load a differential expression analysis.
}
