% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{DEset}
\alias{DEset}
\title{DEset}
\usage{
DEset(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class}
}
\value{
A dataframe with the differential expression data.
}
\description{
Extract the optionally provided differential expression data
}
\details{
\code{DEset} allows one to extract the additional and optional differential expression data.
}
