% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{downloadAE}
\alias{downloadAE}
\title{Download expression data.}
\usage{
downloadAE(AEcodes, path = "data.downloads", processed_data = F)
}
\arguments{
\item{AEcodes}{A vector of ArrayExpress accession codes.}

\item{path}{Path of the directory where the downloaded expression data should be stored. In case this directory does not yet exist, it will be created automatically.}

\item{processed_data}{Logical indicating whether or not to retrieve pre-processed expression data in case the raw data is missing. Defaults to FALSE.}
}
\value{
None.
}
\description{
\code{downloadAE} downloads the expression data associated with provided ArrayExpress accession codes.
}
\examples{
downloadAE(AEcodes = c("E-MTAB-4025","E-MTAB-2495"), path="data.downloads")
}
