% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{expressionData}
\alias{expressionData}
\title{expressionData}
\usage{
expressionData(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
The gene expression matrix used for CenTFinder analysis.
}
\description{
Extract the gene expression matrix
}
\details{
\code{expressionData} allows one to extract the gene expression matrix used for analysis (i.e. any filtering will be taken into account).
}
