% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_integration.R
\name{extractNetwork}
\alias{extractNetwork}
\title{Extract network for Cytoscape visualisation}
\usage{
extractNetwork(
  CenTFinderObject,
  modules = NULL,
  class = c(1, 2, 3),
  DEset = NULL,
  TFs_diff_expr = T,
  Nr_diff_targets = 20,
  target_cell = NULL,
  z_score_cutoff = 2,
  p_value_cutoff = 0.05
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{modules}{Modules of interest.}

\item{class}{Integers specifying the transcription factor classes to be displayed. Class 1 transcription factors are differentially expressed, are WGCNA cluster-central, and have RcisTarget enriched binding motifs in the promoters of the genes of their respective clusters. Can only be 1, 2, and/or 3.}

\item{DEset}{Dataset from which to use the fold changes.}

\item{TFs_diff_expr}{Logical indicating whether or not differentially expressed transcription factors of the module are selected. Transcription factors that are not differentially expressed might play a posttranscriptional role. Defaults to TRUE.}

\item{Nr_diff_targets}{Integer specifying the maximum number of top differentially expressed targets are included in the Cytoscape network. Defaults to 20.}

\item{target_cell}{Cell type in which the TFs and/or targets should be differentially expressed. If target_cell and additional_DE are NULL, TFs_diff_expr and Nr_diff_targets will be ignored.}

\item{z_score_cutoff}{In case target_cell is NULL and additional_DE is FALSE and Z-scores were calculated for the expression data, then this parameter specifies the cutoff to apply.}

\item{p_value_cutoff}{In case target_cell is NULL and additional_DE is FALSE and t-tests were calculated for the expression data, then this parameter specifies the cutoff to apply.}
}
\value{
A dataframe with the desired Cytoscape network
}
\description{
\code{extractNetwork} extracts Cytoscape networks from a CenTFinder object for specified modules.
}
