% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{filterAEdata}
\alias{filterAEdata}
\title{Unpack and filter downloaded expression data.}
\usage{
filterAEdata(
  sdrf,
  path = "data.downloads",
  species = "Homo sapiens",
  keywords = NULL,
  exclusion_keywords = NULL,
  allowed_extensions = c(".cel", ".txt")
)
}
\arguments{
\item{sdrf}{The SDRF data describing the expression data in the provided path. This should be the output of \code{\link{getBatchSDRF}}.}

\item{path}{Path to directory where expression data is stored.}

\item{species}{Desired species of the expression data. Latin names are supported (e.g. 'Homo sapiens').}

\item{keywords}{A character vector containing keywords for files of interest. Expression data for which at least one of the annotation fields contains at least one keyword are kept.}

\item{exclusion_keywords}{A character vector containing keywords for expression data that (for user-specificshould be excluded from analysis (e.g. cancer data). Defaults to NULL.}

\item{allowed_extensions}{A character vector containing the extensions of data types allowed for analysis.}
}
\description{
\code{filterAEdata} unzips all zip files present in the path provided in \code{\link{downloadAE}}. Only expression data of the desired species are retained. Subsequently, the expression data is filtered for keywords.
}
\examples{
sdrf<-getBatchSDRF(AEcodes=c("E-MTAB-4025","E-MTAB-2495"), technology = "array", label = "biotin")
filterAEdata(sdrf = sdrf, path="data.downloads",species = "Homo sapiens", keywords = keywords,exclusion_keywords=c("oma","tumor","tumour","cancer","malignant"))
}
