% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter genes.}
\usage{
filterGenes(
  CenTFinderObject,
  tfs = NULL,
  TF_range_cutoff = 1,
  max_genes = 9000
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{tfs}{A vector of transcription factor gene symbols.}

\item{TF_range_cutoff}{An numeric specifying the minimum log2 fold change of a transcription factor between any two samples. Defaults to 1.}

\item{max_genes}{An integer specifying the maximum total number of genes to be retained for analysis. Increasing this parameter will increase computation time. Computational complexity is of the order O(n^2). Defaults to 9000.}
}
\value{
A CenTFinderObject with a filtered expression matrix.
}
\description{
\code{filterGenes} filters a CenTFinder expression matrix for the most variable genes.
}
\details{
This function applies different filtering criteria for transcription factors and the other genes. Transcription factors are typically retained if they vary more than two-fold across all samples. This behaviour can be regulated with the 'TF_range_cutoff' argument. The most variant genes that are not transcriptional regulators are retained.
}
\examples{
humantfs<-as.character(read.table("allhumantfs.txt",sep = "\t")[,1])
CenTFinderObject<-filterGenes(CenTFinderObject,tfs=humantfs, TF_range_cutoff=1, max_genes=9000)
}
