% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{geneAnnotations}
\alias{geneAnnotations}
\title{geneAnnotations}
\usage{
geneAnnotations(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A dataframe with the gene cluster annotation.
}
\description{
Extract the gene-module annotations
}
\details{
\code{geneAnnotations} allows one to extract the gene-module annotations from a CenTFinder object.
}
