% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{getAE_accession_codes}
\alias{getAE_accession_codes}
\title{getAE_accession_codes}
\usage{
getAE_accession_codes(query)
}
\arguments{
\item{query}{One or more keywords for querying the ArrayExpress database.}
}
\value{
A vector containing the retrieved ArrayExpress accession codes.
}
\description{
getAE_accession_codes
}
\examples{
getAE_accession_codes(c("endothelial cell AND Homo sapiens","LSEC AND Homo sapiens"))
}
