% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{getAllVariances}
\alias{getAllVariances}
\title{getAllVariances}
\usage{
getAllVariances(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A dataframe with the variance of each gene.
}
\description{
Extract the variances of all genes
}
\details{
\code{getAllVariances} extracts all gene expression variances.
}
