% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{getBatchSDRF}
\alias{getBatchSDRF}
\title{Get sample-data relationship data.}
\usage{
getBatchSDRF(
  AEcodes,
  technology = c("array", "sequencing"),
  label = if (technology == "array") {     "biotin" } else {     NULL },
  cores = detectCores() - 1
)
}
\arguments{
\item{AEcodes}{A vector containing ArrayExpress accession codes.}

\item{technology}{Has to be either "array" or "sequencing".}

\item{label}{Is NULL in case 'technology = sequencing', otherwise it defaults to 'biotin'. Here, different array types can be specified (i.e. biotin arrays, dual arrays, ...).}

\item{cores}{Integer specifying the number of cores to be used. Defaults to the number of available cores minus one.}
}
\value{
A dataframe containing the merged SDRF files.
}
\description{
\code{getBatchSDRF} retrieves the ArrayExpress Sample Data Relationships Files (SDRF) for the provided accession codes.
}
\details{
After using \code{\link{getAE_accession_codes}} this function should be used to retrieve the SDRF data of the acquired accession codes. This function further filters the datasets for the desired assay type (i.e. sequencing or array).
}
\examples{
getBatchSDRF(AEcodes=c("E-MTAB-4025","E-MTAB-2495"), technology = "array", label = "biotin")
}
