% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{getModuleScores}
\alias{getModuleScores}
\title{Extract the GSVA module scores of the CenTFinder object}
\usage{
getModuleScores(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class}
}
\value{
A dataframe with containing the GSVA activity
}
\description{
\code{getModuleScores} lets one extract the GSVA module activities from a CenTFinder object.
}
