% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{getOverlap}
\alias{getOverlap}
\title{Extract the overlap between WGCNA, RcisTarget, and any optional differential expression sets}
\usage{
getOverlap(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A dataframe containing the overlap of all analyses.
}
\description{
\code{getOverlap} extracts the overlap or integration of all analyses.
}
