% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{getPlatformInfo}
\alias{getPlatformInfo}
\title{Gather information about the downloaded platforms.}
\usage{
getPlatformInfo(
  CenTFinderObject,
  filenames = NULL,
  path,
  platform_freq_cutoff = 0
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{path}{The path pointing to the downloaded CEL files.}

\item{platform_freq_cutoff}{Minimal percentage of arrays allowed of any given platform. Defaults to zero.}
}
\value{
A CenTFinderObject with updated platform annotation.
}
\description{
\code{getPlatformInfo} evaluates the platform type and frequency of the downloaded expression data.
}
\examples{
CenTFinderObject<-getPlatformInfo(CenTFinderObject,path = "data.downloads", platform_freq_cutoff = 0.00258)
}
