% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Extract the sample names}
\usage{
getSampleNames(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class}
}
\value{
A character vector with the sample names used.
}
\description{
\code{getSampleNames} allows to extract the sample names of samples used in the CenTFinder analysis.
}
