% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{getVariableImportances}
\alias{getVariableImportances}
\title{Extract the variable importances for each WWGCNA module}
\usage{
getVariableImportances(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A dataframe with the variable importance of each module.
}
\description{
\code{getVariableImportances} extracts the variable importance for each module. This is a wrapper function for the 'filterVarImp' function of the 'caret' package.
}
