% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{groups}
\alias{groups}
\title{Extract groups}
\usage{
groups(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A vector containing all unique group names.
}
\description{
\code{groups} extracts the groups within a CenTFinder object.
}
