% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_integration.R
\name{integrateAnalyses}
\alias{integrateAnalyses}
\title{integrateAnalyses}
\usage{
integrateAnalyses(CenTFinderObject, tfs = NULL, kME_cutoff = 0.7)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{tfs}{A vector of transcription factor gene symbols.}

\item{kME_cutoff}{A numeric cut-off for cluster centrality. Cluster centrality is measured with the module eigengene. A kME of 1 indicates complete centrality while a kME of 0 indicates no cluster membership.}
}
\value{
An updated CenTFinderObject with added integrative summary, module GSVA scores, and module variable importances.
}
\description{
integrateAnalyses
}
\examples{
CenTFinderObject<-integrateAnalyses(CenTFinderObject, tfs=humantfs, target_cell = "LSEC")
}
