% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{loadExpression}
\alias{loadExpression}
\title{Load expression matrix.}
\usage{
loadExpression(
  CenTFinderObject,
  expr,
  colData,
  ID_type = c("SYMBOL", "ENSEMBL"),
  species = c("Homo sapiens", "Mus musculus"),
  statistic = c("Wilcoxon", "T-test", "Zscores", "none"),
  contrast = NULL
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CeNTFinder class.}

\item{expr}{The expression matrix, either microarray, bulkRNAseq, or TempO-Seq data.}

\item{colData}{An annotation dataframe for the samples. Should have two columns: File and Group. File names should match the column names in the expression data.}

\item{ID_type}{Type of gene identifier. Currently, 'SYMBOL' and 'ENSEMBL' ID types are supported.}

\item{species}{Species of the samples. Currently, only 'Homo sapiens' and 'Mus musculus' are supported.}
}
\value{
A CenTFinder object updated with the expression matrix.
}
\description{
\code{loadExpression} loads a clean expression matrix into a CenTFinder object. Expression values can be log2(intensity) values from microarray data, log2(TPM+1) counts from bulk RNA sequencing data or log2(CPM+1) values from TempO-Seq data.
}
