% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{loadProbeAnnotation}
\alias{loadProbeAnnotation}
\title{Add platform probe annotation data.}
\usage{
loadProbeAnnotation(
  CenTFinderObject,
  platform = NULL,
  probes = NULL,
  symbols = NULL
)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{platform}{Name of the platform.}

\item{probes}{A vector of probe identifiers.}

\item{symbols}{A vector of gene symbols corresponding to the respective probes.}
}
\value{
A CenTFinderObject with additional probe annotation data.
}
\description{
\code{loadProbeAnnotation} allows to add probe annotation for platforms other than HG-U133_Plus_2, NuGO_Hs1a520180 , HG-U219, or HT_HG-U133A. Arrays not belonging to aforementioned four platform types and for which no probe annotation is provided through \code{loadProbeAnnotation} will be excluded from further analysis.
}
