% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{modules}
\alias{modules}
\title{Extract the module colours}
\usage{
modules(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A character vector containing the module colours in the CenTFinder analysis.
}
\description{
\code{modules} allows one to extract the module colours of a CenTFinder object.
}
