% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{platforms}
\alias{platforms}
\title{Extract microarray platforms}
\usage{
platforms(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A character vector containing the platform types used.
}
\description{
\code{platforms} returns the microarray platform types that were used in the CenTFinder analysis.
}
