% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,CenTFinder,ANY-method}
\alias{plot,CenTFinder,ANY-method}
\title{Plot an aspect of a CenTFinder object}
\usage{
\S4method{plot}{CenTFinder,ANY}(x, aspect, ...)
}
\arguments{
\item{aspect}{An aspect of the CenTFinder class. Aspect can be either of the following: "array distribution" (plots the microarray platform distribution), "group distribution" (plots the distribution of the different groups (cell types)), "expression" (plots the expression of one or more genes in the meta-analysis data), "markers" (plots the expression of top identified markers in the meta-analysis data), "tfs" (plots the expression of top expressed tfs in the cell type of interest in the meta-analysis data), "variance" (plots the least and most variable genes in the meta-analysis data), "scale independence" (plots the correlation of each soft-thresholding connectivity distribution with the scale-free topology connectivity distribution), "mean connectivity" (plots the mean connectivity in the network in function of the soft-thresholding power used), "tree" (plots the WGCNA dendrogram tree), "cluster dendrogram" (plots the WGCNA cluster dendrogram), "GSVA heatmap" (plots a heatmap of GSVA activity scores for each module and each cell type/group of cells), "GSVA boxplot" (plots a boxplot of the GSVA scores of the cell type/group of interest versus the other cell types/groups), "module PCA" (plots a PCA plot for a given module), "GO" (plots the gene ontology enerichment for a given module), "KEGG" (plots the KEGG enrichment for a given module),"RcisTarget" (plots the RcisTarget results), and "integrative" (plots the overlap and final ranking of all analyses).}

\item{object}{An instance of the CenTFinder class.}
}
\description{
Plot allows one to visualise the different aspects of the CenTFinder analysis.
}
