% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{removeSamples}
\alias{removeSamples}
\title{Remove samples}
\usage{
removeSamples(CenTFinderObject, samples = NULL)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{samples}{Names or indices of the samples to be removed.}
}
\value{
An updated CenTFinder object in which samples were removed.
}
\description{
\code{removeSamples} allows one to remove samples from a CenTFinder object.
}
