% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{sampleAnnotations}
\alias{sampleAnnotations}
\title{Extract sample annotation}
\usage{
sampleAnnotations(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
A dataframe containing the sample annotation of the samples in a CenTFinder object.
}
\description{
\code{sampleAnnotations} extracts the sample annotation of a CenTFinder object.
}
