% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions_preprocessing.R
\name{setArrayAnnotations}
\alias{setArrayAnnotations}
\title{Set array annotations.}
\usage{
setArrayAnnotations(CenTFinderObject, files, groups, AEcodes)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}

\item{files}{A vector of CEL file names in the download path.}

\item{groups}{Groups corresponding to the provided CEL file names. Groups could indicate the cell type, but any other grouping is possible.}

\item{AEcodes}{Accession codes corresponding to the provided CEL file names.}
}
\value{
A CenTFinderObject with updated sample annotations.
}
\description{
\code{setArrayAnnotations} sets for each sample the provided cell type and accession code. This information is easily put together based on the SDRF information.
}
