% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,CenTFinder-method}
\alias{show,CenTFinder-method}
\title{Show a summary of a CenTFinder object}
\usage{
\S4method{show}{CenTFinder}(object)
}
\arguments{
\item{object}{An instance of the CenTFinder class.}
}
\description{
Show reveals the main features of the CenTFinder object, including analyses performed.
}
