% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,CenTFinder-method}
\alias{summary,CenTFinder-method}
\title{Get a summary of a CenTFinder object}
\usage{
\S4method{summary}{CenTFinder}(object, ...)
}
\arguments{
\item{object}{An instance of the CenTFinder class.}
}
\description{
Summary reveals the main features of the CenTFinder object, including analyses performed.
}
