% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_misc.R
\name{thresholdingPower}
\alias{thresholdingPower}
\title{Extract thresholding power}
\usage{
thresholdingPower(CenTFinderObject)
}
\arguments{
\item{CenTFinderObject}{An instance of the CenTFinder class.}
}
\value{
The applied soft-thresholding power.
}
\description{
\code{tresholdingPower} extracts the soft-thresholding power used in the CenTFinder analysis.
}
